(#14 #15) Nice Amount View Columns
Moved the responsibility of calling the column painter into the
ItemDelegate because the ItemDelegate has information about the
decorations which are painted.  The ItemDelegate class must instruct
the value painter where to paint it's value by passing the appropriate
option.rect.

When a new font was selected in the preferences panel, the column would
redraw but not resize appropriately.  A call to resize(sizeHint()) was
added on the update of the font size.

The column painter now paints total rows, and a much simpler check of the
column.painter attribute against None replaces the need for that hack line
in the _getData method of the base table class in qtlib.