(#26) Intelligent Import Autofill
Fixed all tests, missed some logic regarding the filtering of panes
to only those able to perform an action.  Also, some of the date-names
were not being set until after the "selected pane" update method was
called.  The user would not notice it, but the tests expected the name
to be set more immediately.