(#391) Getting toward "ideal" budget spawn record
Subclass _Transaction so that a spawn can record some values necessary
for the budget.  I don't think this approach will be permanent, but I
needed a spot for these records that still worked with recurrences.

Rather than keeping a separate association of the spawn / spawn data in
```.Budget``` I thought it would be better to put that in the spawn
itself.  Some thoughts are in comments.